Implement error handling on order routes

+++ services/Order.service.js
@@ -1,4 +1,9 @@
 import OrderModel from "../models/Order.model.js"
+import {
+  BadRequestError,
+  NotFoundError,
+  ValidationError,
+} from "../utils/customErrorHandler.js"
 
 export const getAllOrders = async (req, res) => {
   try {
@@ -12,7 +17,13 @@ export const getAllOrders = async (req, res) => {
 
 export const getOrderByUserId = async (req, res) => {
   try {
+    if (!req.params.id) {
+      throw new BadRequestError("Not mentioned user id on API.")
+    }
     const allOrders = await OrderModel.find({ userId: req.params.id })
+    if (allOrders.length === 0) {
+      throw new NotFoundError("No order found.")
+    }
     res.status(200)
     res.json(allOrders)
   } catch (error) {
@@ -22,7 +33,13 @@ export const getOrderByUserId = async (req, res) => {
 
 export const getOrderByOrderId = async (req, res) => {
   try {
+    if (!req.params.id) {
+      throw new BadRequestError("Not mentioned order id on API.")
+    }
     const order = await OrderModel.find({ id: req.params.id })
+    if (order.length === 0) {
+      throw new NotFoundError("No order found.")
+    }
     res.status(200)
     res.json(order)
   } catch (error) {
@@ -32,20 +49,40 @@ export const getOrderByOrderId = async (req, res) => {
 
 export const saveNewOrder = async (req, res) => {
   try {
+    if (!req.body) {
+      throw new BadRequestError("Request body is missing.")
+    }
     const NewOrder = new OrderModel(req.body)
     await NewOrder.save()
     res.status(200)
     res.json(NewOrder)
   } catch (error) {
-    throw error
+    if (error.name === "ValidationError") {
+      throw new ValidationError(error.message)
+    } else if (error.code === 11000) {
+      const field = Object.keys(error.keyPattern)[0]
+      throw new ValidationError(`${field} must be unique`)
+    } else {
+      throw error
+    }
   }
 }
 
 export const findOrderByIdAndUpdate = async (req, res) => {
   try {
     const id = req.params.id
+    if (!id) {
+      throw new BadRequestError("Order id is missing.")
+    }
+    const Order = await OrderModel.find({ id })
+    if (Order.length === 0) {
+      throw new NotFoundError("Order not found.")
+    }
     const dataToUpdate = req.body
-    const order = await OrderModel.findOneAndUpdate({ id: id }, dataToUpdate, {
+    if (!dataToUpdate) {
+      throw new BadRequestError("Request body is missing.")
+    }
+    const order = await OrderModel.findOneAndUpdate({ id }, dataToUpdate, {
       new: true,
     })
     res.status(200)
@@ -57,6 +94,13 @@ export const findOrderByIdAndUpdate = async (req, res) => {
 
 export const findOrderByIdAndDelete = async (req, res) => {
   try {
+    if (!req.params.id) {
+      throw new BadRequestError("Order id is missing.")
+    }
+    const Order = await OrderModel.find({ id: req.params.id })
+    if (Order.length === 0) {
+      throw new NotFoundError("Order not found.")
+    }
     const deletedOrder = await OrderModel.findOneAndDelete({
       id: req.params.id,
     })
